Disable copying of file objects
Copy constructor of file objects were not deleted,
which led to a subtle bug, where a file was copied,
and then doubly closed, which crashed the system.
For that reason the destructor of that class did nothing.

This change disables copying of file objects,
defines appropriate move constructor and assignment
operator, as well as fixes the destructor and the
original bug.